+++ XamarinApp/LAMA/LAMA/LAMA/database/RememberedList.cs
@@ -12,7 +12,7 @@ namespace LAMA
     
     public class RememberedList<T, Storage> where T : Serializable, new() where Storage : Database.StorageInterface, new()
     {
-        SortedDictionary<int, int> IDToIndex = new SortedDictionary<int, int>();
+        SortedDictionary<long, int> IDToIndex = new SortedDictionary<long, int>();
 
         
         List<T> cache = new List<T>();
@@ -99,7 +99,7 @@ namespace LAMA
         }
 
         //There was no way for me to do this with the original interface (I couldn't get the item index based on ID)
-        public void removeByID(int ID)
+        public void removeByID(long ID)
         {
             if (IDToIndex.ContainsKey(ID))
             {
@@ -108,7 +108,7 @@ namespace LAMA
             }
         }
 
-        public T getByID(int id)
+        public T getByID(long id)
         {
             if (!IDToIndex.ContainsKey(id))
                 return default(T);
@@ -137,7 +137,7 @@ namespace LAMA
         /// 
         /// </summary>
         /// <returns> returns -1 if it is waiting for more ID intervals from the server </returns>
-        public int nextID()
+        public long nextID()
         {   
 
             Database.Interval currentInterval = null;
